Surprise event raiser refactorisation
It's fine putting SerializeField tags in Ancestor classes, but the respective field(s) need(s) to be set in the child class afterwards in the editor.

Also, the Untiy methods need to be, not overridden, but superseded by dobbelgänger methods which are invoked in the next layer
e.g. _onTriggerEnter2D of the ancestor guy must be invoked in the child class OnTriggerExit2D